src/backend/requirements.txt
aiobotocore==2.25.1
aioduckdb==0.1.0
aiohttp==3.13.2
aiosmtplib==5.0.0
aiosqlite==0.21.0
argon2-cffi==25.1.0
asgi-ratelimit==0.10.0
bcrypt==5.0.0
debugpy==1.8.17
msgspec==0.19.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation-logging==0.59b0
opentelemetry-instrumentation-starlette==0.59b0
starlette==0.50.0
types-aiobotocore[s3,ses]==2.25.1
uvicorn==0.38.0